Arm backend: Move get_tosa_spec to tosa_specification.py
